PLUGIN-87. Salesforce streaming plugin validate button is broken
