+++ src/main/java/io/cdap/plugin/salesforce/plugin/source/streaming/SalesforceStreamingSource.java
@@ -24,7 +24,6 @@ import io.cdap.cdap.api.annotation.Description;
 import io.cdap.cdap.api.annotation.Name;
 import io.cdap.cdap.api.annotation.Plugin;
 import io.cdap.cdap.api.data.format.StructuredRecord;
-import io.cdap.cdap.api.data.format.UnexpectedFormatException;
 import io.cdap.cdap.api.data.schema.Schema;
 import io.cdap.cdap.api.dataset.DatasetProperties;
 import io.cdap.cdap.etl.api.PipelineConfigurer;
@@ -37,20 +36,10 @@ import io.cdap.plugin.salesforce.SObjectDescriptor;
 import io.cdap.plugin.salesforce.SalesforceSchemaUtil;
 import io.cdap.plugin.salesforce.authenticator.Authenticator;
 import io.cdap.plugin.salesforce.authenticator.AuthenticatorCredentials;
-import org.apache.spark.api.java.function.Function;
 import org.apache.spark.streaming.api.java.JavaDStream;
-import org.apache.spark.streaming.api.java.JavaStreamingContext;
-import org.json.JSONException;
-import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.time.Instant;
-import java.time.LocalTime;
-import java.time.temporal.ChronoUnit;
-import java.util.Map;
-import java.util.Objects;
-import java.util.concurrent.TimeUnit;
 import javax.ws.rs.Path;
 
 /**
@@ -110,83 +99,9 @@ public class SalesforceStreamingSource extends StreamingSource<StructuredRecord>
     }
     LOG.debug("Schema is {}", schema);
 
-    JavaStreamingContext jssc = streamingContext.getSparkStreamingContext();
+    SalesforceStreamer streamer = new SalesforceStreamer(schema, config);
 
-    return jssc.receiverStream(new SalesforceReceiver(this.config.getAuthenticatorCredentials(),
-                                                      this.config.getPushTopicName())).
-      map((Function<String, StructuredRecord>) this::getStructuredRecord).filter(Objects::nonNull);
-  }
-
-  private StructuredRecord getStructuredRecord(String jsonMessage) {
-    StructuredRecord.Builder builder = StructuredRecord.builder(schema);
-
-    JSONObject sObjectFields;
-    try {
-      sObjectFields = new JSONObject(jsonMessage) // throws a JSONException if failed to decode
-        .getJSONObject("sobject"); // throws a JSONException if not found
-    } catch (JSONException e) {
-      throw new IllegalStateException(
-        String.format("Cannot retrieve /data/sobject from json message %s", jsonMessage), e);
-    }
-
-    for (Map.Entry<String, Object> entry : sObjectFields.toMap().entrySet()) {
-      String fieldName = entry.getKey();
-      Object value = entry.getValue();
-
-      Schema.Field field = schema.getField(fieldName, true);
-
-      if (field == null) {
-        continue; // this field is not in schema
-      }
-
-      builder.set(field.getName(), convertValue(value, field));
-    }
-    return builder.build();
-  }
-
-  private Object convertValue(Object value, Schema.Field field) {
-    if (value == null) {
-      return null;
-    }
-
-    Schema fieldSchema = field.getSchema();
-
-    if (fieldSchema.isNullable()) {
-      fieldSchema = fieldSchema.getNonNullable();
-    }
-
-    Schema.Type fieldSchemaType = fieldSchema.getType();
-    Schema.LogicalType logicalType = fieldSchema.getLogicalType();
-
-    if (fieldSchema.getLogicalType() != null) {
-      String valueString = (String) value;
-      switch (logicalType) {
-        case DATE:
-          return Math.toIntExact(ChronoUnit.DAYS.between(Instant.EPOCH, Instant.parse(valueString)));
-        case TIMESTAMP_MICROS:
-          return TimeUnit.MILLISECONDS.toMicros(Instant.parse(valueString).toEpochMilli());
-        case TIME_MICROS:
-          return TimeUnit.NANOSECONDS.toMicros(LocalTime.parse(valueString).toNanoOfDay());
-        default:
-          throw new UnexpectedFormatException(String.format("Field '%s' is of unsupported type '%s'",
-                                                            field.getName(), logicalType.getToken()));
-      }
-    }
-
-    // Found a single field (Opportunity.Fiscal) which is documented as string and has a string type
-    // in describe result, however in Salesforce Streaming API reponse json is represented as json.
-    // Converting it and similar back to string, since it does not comply with generated schema.
-    if (value instanceof Map) {
-      if (fieldSchemaType.equals(Schema.Type.STRING)) {
-        return value.toString();
-      } else {
-        throw new UnexpectedFormatException(
-          String.format("Field '%s' is of type '%s', but value found is '%s'",
-                        field.getName(), fieldSchemaType.toString(), value.toString()));
-      }
-    }
-
-    return value;
+    return streamer.getReceiverStream(streamingContext);
   }
 
   @Path("outputSchema")
